Move UserDetailsRepository to core.userdetails
Fixes gh-4383
